refactor category and tags
