+++ modules/config.py
@@ -474,240 +474,6 @@ class Config:
             self.notify(e, "Config")
             raise Failed(e)
 
-    # Get tags from config file based on keyword
-    def get_tags(self, trackers):
-        urls = [x.url for x in trackers if x.url.startswith("http")]
-        tracker = {}
-        tracker["tag"] = None
-        tracker["max_ratio"] = None
-        tracker["min_seeding_time"] = None
-        tracker["max_seeding_time"] = None
-        tracker["limit_upload_speed"] = None
-        tracker["notifiarr"] = None
-        tracker["url"] = None
-        tracker_other_tag = self.util.check_for_attribute(
-            self.data, "tag", parent="tracker", subparent="other", default_is_none=True, var_type="list", save=False
-        )
-        try:
-            tracker["url"] = util.trunc_val(urls[0], os.sep)
-        except IndexError as e:
-            tracker["url"] = None
-            if not urls:
-                urls = []
-                if not tracker_other_tag:
-                    tracker_other_tag = ["other"]
-                tracker["url"] = "No http URL found"
-            else:
-                logger.debug(f"Tracker Url:{urls}")
-                logger.debug(e)
-        if "tracker" in self.data and self.data["tracker"] is not None:
-            tag_values = self.data["tracker"]
-            for tag_url, tag_details in tag_values.items():
-                for url in urls:
-                    if tag_url in url:
-                        if tracker["url"] is None:
-                            default_tag = tracker_other_tag
-                        else:
-                            try:
-                                tracker["url"] = util.trunc_val(url, os.sep)
-                                default_tag = tracker["url"].split(os.sep)[2].split(":")[0]
-                            except IndexError as e:
-                                logger.debug(f"Tracker Url:{url}")
-                                logger.debug(e)
-                        # Tracker Format 1 deprecated.
-                        if isinstance(tag_details, str):
-                            e = (
-                                "Config Error: Tracker format invalid. Please see config.yml.sample for correct format and fix "
-                                f"`{tag_details}` in the Tracker section of the config."
-                            )
-                            self.notify(e, "Config")
-                            raise Failed(e)
-                        # Using new Format
-                        else:
-                            tracker["tag"] = self.util.check_for_attribute(
-                                self.data, "tag", parent="tracker", subparent=tag_url, default=tag_url, var_type="list"
-                            )
-                            if tracker["tag"] == [tag_url]:
-                                self.data["tracker"][tag_url]["tag"] = [tag_url]
-                            if isinstance(tracker["tag"], str):
-                                tracker["tag"] = [tracker["tag"]]
-                            is_max_ratio_defined = self.data["tracker"].get("max_ratio")
-                            is_max_seeding_time_defined = self.data["tracker"].get("max_seeding_time")
-                            if is_max_ratio_defined or is_max_seeding_time_defined:
-                                tracker["max_ratio"] = self.util.check_for_attribute(
-                                    self.data,
-                                    "max_ratio",
-                                    parent="tracker",
-                                    subparent=tag_url,
-                                    var_type="float",
-                                    min_int=-2,
-                                    do_print=False,
-                                    default=-1,
-                                    save=False,
-                                )
-                                tracker["max_seeding_time"] = self.util.check_for_attribute(
-                                    self.data,
-                                    "max_seeding_time",
-                                    parent="tracker",
-                                    subparent=tag_url,
-                                    var_type="int",
-                                    min_int=-2,
-                                    do_print=False,
-                                    default=-1,
-                                    save=False,
-                                )
-                            else:
-                                tracker["max_ratio"] = self.util.check_for_attribute(
-                                    self.data,
-                                    "max_ratio",
-                                    parent="tracker",
-                                    subparent=tag_url,
-                                    var_type="float",
-                                    min_int=-2,
-                                    do_print=False,
-                                    default_is_none=True,
-                                    save=False,
-                                )
-                                tracker["max_seeding_time"] = self.util.check_for_attribute(
-                                    self.data,
-                                    "max_seeding_time",
-                                    parent="tracker",
-                                    subparent=tag_url,
-                                    var_type="int",
-                                    min_int=-2,
-                                    do_print=False,
-                                    default_is_none=True,
-                                    save=False,
-                                )
-                            tracker["min_seeding_time"] = self.util.check_for_attribute(
-                                self.data,
-                                "min_seeding_time",
-                                parent="tracker",
-                                subparent=tag_url,
-                                var_type="int",
-                                min_int=0,
-                                do_print=False,
-                                default=0,
-                                save=False,
-                            )
-                            tracker["limit_upload_speed"] = self.util.check_for_attribute(
-                                self.data,
-                                "limit_upload_speed",
-                                parent="tracker",
-                                subparent=tag_url,
-                                var_type="int",
-                                min_int=-1,
-                                do_print=False,
-                                default=0,
-                                save=False,
-                            )
-                            tracker["notifiarr"] = self.util.check_for_attribute(
-                                self.data,
-                                "notifiarr",
-                                parent="tracker",
-                                subparent=tag_url,
-                                default_is_none=True,
-                                do_print=False,
-                                save=False,
-                            )
-                        return tracker
-            if tracker_other_tag:
-                tracker["tag"] = tracker_other_tag
-                tracker["max_ratio"] = self.util.check_for_attribute(
-                    self.data,
-                    "max_ratio",
-                    parent="tracker",
-                    subparent="other",
-                    var_type="float",
-                    min_int=-2,
-                    do_print=False,
-                    default=-1,
-                    save=False,
-                )
-                tracker["min_seeding_time"] = self.util.check_for_attribute(
-                    self.data,
-                    "min_seeding_time",
-                    parent="tracker",
-                    subparent="other",
-                    var_type="int",
-                    min_int=0,
-                    do_print=False,
-                    default=-1,
-                    save=False,
-                )
-                tracker["max_seeding_time"] = self.util.check_for_attribute(
-                    self.data,
-                    "max_seeding_time",
-                    parent="tracker",
-                    subparent="other",
-                    var_type="int",
-                    min_int=-2,
-                    do_print=False,
-                    default=-1,
-                    save=False,
-                )
-                tracker["limit_upload_speed"] = self.util.check_for_attribute(
-                    self.data,
-                    "limit_upload_speed",
-                    parent="tracker",
-                    subparent="other",
-                    var_type="int",
-                    min_int=-1,
-                    do_print=False,
-                    default=0,
-                    save=False,
-                )
-                tracker["notifiarr"] = self.util.check_for_attribute(
-                    self.data,
-                    "notifiarr",
-                    parent="tracker",
-                    subparent="other",
-                    default_is_none=True,
-                    do_print=False,
-                    save=False,
-                )
-                return tracker
-        if tracker["url"]:
-            logger.trace(f"tracker url: {tracker['url']}")
-            if tracker_other_tag:
-                default_tag = tracker_other_tag
-            else:
-                default_tag = tracker["url"].split(os.sep)[2].split(":")[0]
-            tracker["tag"] = self.util.check_for_attribute(
-                self.data, "tag", parent="tracker", subparent=default_tag, default=default_tag, var_type="list"
-            )
-            if isinstance(tracker["tag"], str):
-                tracker["tag"] = [tracker["tag"]]
-            try:
-                self.data["tracker"][default_tag]["tag"] = [default_tag]
-            except Exception:
-                self.data["tracker"][default_tag] = {"tag": [default_tag]}
-            e = f'No tags matched for {tracker["url"]}. Please check your config.yml file. Setting tag to {default_tag}'
-            self.notify(e, "Tag", False)
-            logger.warning(e)
-        return tracker
-
-    # Get category from config file based on path provided
-    def get_category(self, path):
-        category = ""
-        path = os.path.join(path, "")
-        if "cat" in self.data and self.data["cat"] is not None:
-            cat_path = self.data["cat"]
-            for cat, save_path in cat_path.items():
-                if os.path.join(save_path, "") == path:
-                    category = cat
-                    break
-
-        if not category:
-            default_cat = path.split(os.sep)[-2]
-            category = str(default_cat)
-            self.util.check_for_attribute(self.data, default_cat, parent="cat", default=path)
-            self.data["cat"][str(default_cat)] = path
-            e = f"No categories matched for the save path {path}. Check your config.yml file. - Setting category to {default_cat}"
-            self.notify(e, "Category", False)
-            logger.warning(e)
-        return category
-
     # Empty old files from recycle bin or orphaned
     def cleanup_dirs(self, location):
         num_del = 0
