+++ modules/core/__init__.py
@@ -1,0 +1,3 @@
+"""
+ modules.core contains all the core functions of qbit_manage such as updating categories/tags etc..
+"""
